Added new routes

+++ backend/app/controllers/users_controller.rb
@@ -1,21 +1,51 @@
 class UsersController < ApplicationController
 
+  def accept
+    user1 = User.find_by_id(params[:one])
+    user2 = User.find_by_id(params[:two])
+    loc1 = Location.find_by_user_id(params[:one])
+    loc2 = Location.find_by_user_id(params[:two])
+    response = 0
+
+    if !user1.nil? and !loc1.nil? and !user2.nil? and !loc2.nil? and user1.partner == 0 and user2.partner == 0
+      user1.partner = -2 
+      user2.partner = -2
+      user1.save 
+      user2.save
+      response = 1
+    end
+   
+    respond_to do |format|
+      format.json { render :json => response }
+    end
+  end
+
+  def reject
+    user1 = User.find_by_id(params[:one])
+    user2 = User.find_by_id(params[:two])
+    loc1 = Location.find_by_user_id(params[:one])
+    loc2 = Location.find_by_user_id(params[:two])
+    response = 0
+
+    if !user1.nil? and !loc1.nil? and !user2.nil? and !loc2.nil? and user1.partner == 0 and user2.partner == 0
+      user1.partner = 0 
+      user2.partner = 0 
+      user1.save 
+      user2.save
+      response = 1
+    end
+   
+    respond_to do |format|
+      format.json { render :json => response }
+    end
+  end
+
   def partner
     user = User.find_by_id(params[:id])
-    response = -1
+    response = 0 
 
-    if user.nil?
-      response = -1 
-    else
-      if user.active == 0
-        response = -1 
-      else
-        if user.partner == 0 
-          response = -1
-        else
-          response = user.partner
-        end
-      end
+    if !user.nil? and user.active == 1 and user.partner != 0 
+      response = user.partner
     end
 
     respond_to do |format|
@@ -25,11 +55,9 @@ class UsersController < ApplicationController
 
   def changeinfo
     user = User.find_by_id(params[:id])
-    response = -1
+    response = 0 
 
-    if user.nil?
-      response = 0  
-    else
+    if !user.nil?
       user.name = params[:name]
       user.year = params[:year]
       user.save
@@ -43,22 +71,12 @@ class UsersController < ApplicationController
 
   def login
     user = User.find_by_username(params[:username])
-    response = -1
+    response = 0 
    
-    if user.nil?
-      response = 3 # no such user
-    else
-      if user.active == 1
-        response = 2 # already logged on
-      else
-        if user.password == params[:password] 
-          user.active = 1 
-          user.save
-          response = 1 # success
-        else
-          response = 0 # incorrect password
-        end
-      end
+    if !user.nil? and user.active == 0 and user.password == params[:password] 
+       user.active = 1 
+       user.save
+       response = 1 # success
     end
 
     respond_to do |format|
@@ -68,12 +86,9 @@ class UsersController < ApplicationController
 
   def logout
     user = User.find_by_username(params[:username])
-    response = -1
+    response = 0 
 
-    if user.nil?
-      response = 3 # no such user
-    else
-      if user.active == 1
+    if !user.nil? and user.active == 1
         user.active = 0
         user.location.destroy
         partner = user.partner 
@@ -83,9 +98,6 @@ class UsersController < ApplicationController
         otherUser.save
         user.save
         response = 1 # succesful log out 
-      else
-        response = 0 # not logged in
-      end
     end
   
     respond_to do |format|
@@ -121,60 +133,50 @@ class UsersController < ApplicationController
 
   def sendloc
     user = User.find_by_id(params[:id])
-    response = -1
+    response = 0 
 
-    if !user.nil?
-        if user.active == 0
-          response = 0
-        else 
-          loc = user.location
-          lat = Float(params[:lat])
-          long = Float(params[:long])
-  
-          if(loc.nil?)
-            loc = Location.new(:lat => lat, :long => long)
-            loc.user_id = user.id
-            user.location = loc
-            user.save
-            response = 1
-          else
-            Location.destroy(loc.id)
-            newLoc = Location.new(:lat => lat, :long => long)
-            newLoc.user_id = user.id
-            user.location = newLoc
-            user.save
-            response = 1
-          end
-        end
-
-        respond_to do |format|
-          format.json { render :json => response }
-        end
-    else
-      # user is nil
-      response = 0
-      respond_to do |format|
-          format.json { render :json => response }
+    if !user.nil? and user.active == 1
+      loc = user.location
+      lat = Float(params[:lat])
+      long = Float(params[:long])
+ 
+      if loc.nil?
+        loc = Location.new(:lat => lat, :long => long)
+        loc.user_id = user.id
+        user.location = loc
+        user.save
+        response = 1
+      else
+        Location.destroy(loc.id)
+        newLoc = Location.new(:lat => lat, :long => long)
+        newLoc.user_id = user.id
+        user.location = newLoc
+        user.save
+        response = 1
       end
+    end
+
+    respond_to do |format|
+      format.json { render :json => response }
     end
   end
 
   def closestMatch
     user = User.find_by_id(params[:id])
+    response = 0
+    #response = 'hello' 
 
-    if user.nil? or user.active == 0 
-      respond_to do |format|
-        format.json { render :json => 0 }
-      end
-    else
+    if !user.nil? and user.active == 1 
+      #response = "blah"
       loc = Location.find_by_user_id(params[:id])
 
+      minLoc = 0 
       if !loc.nil?
-        minLoc = 0 
-        activeUsers = User.where(:active => 1)
+        #response = "blah3"
+        activeUsers = User.where("active = 1 AND partner = 0")
         locs = activeUsers.collect{ |user| user.location }
         locs.each { |x| 
-            if x != nil and x.user_id != loc.user_id
+          if x != nil and x.user_id != loc.user_id
             if minLoc == 0 
               minLoc = x
             elsif Math.sqrt((loc.lat - x.lat)**2 + (loc.long - x.long)**2)  < Math.sqrt((loc.lat - minLoc.lat)**2 + (loc.long - minLoc.long)**2) 
@@ -183,22 +185,42 @@ class UsersController < ApplicationController
           end
         }
 
+        #response = 'bye'
         if minLoc != 0
+          #response = 'hola'
           other = User.find_by_id(minLoc.user_id) 
           other.partner = user.id
           user.partner = other.id
           other.save
           user.save
-        end
-        respond_to do |format|
-          format.json { render :json => minLoc }
-        end
-      else
-        respond_to do |format|
-          format.json { render :json => 0 }
+          response = minLoc 
         end
       end
     end 
+  
+    respond_to do |format|
+      format.json { render :json => response }
+    end
+  end
+
+  def chooseMatch
+    user1 = User.find_by_id(params[:one])
+    loc1 = Location.find_by_user_id(params[:one])
+    user2 = User.find_by_id(params[:two])
+    loc2 = Location.find_by_user_id(params[:two])
+    response = 0
+
+    if !user1.nil? and !loc1.nil? and !user2.nil? and !loc2.nil? and user1.partner == 0 and user2.partner == 0
+      user1.partner = Integer(params[:two])
+      user2.partner = Integer(params[:one])
+      user1.save 
+      user2.save
+      response = 1
+    end
+   
+    respond_to do |format|
+      format.json { render :json => response }
+    end
   end
 
   def showusers
@@ -220,20 +242,16 @@ class UsersController < ApplicationController
 
   def create
     oldUser = User.find_by_username(params[:username])
+    response = 0
 
-    if !oldUser.nil?
-      respond_to do |format|
-        format.json { render :json => 0 }
-      end
-    else
+    if oldUser.nil?
       user = User.new(:name => params[:name], :year => params[:year], :username => params[:username], :password => params[:password])
-      respond_to do |format|
-        if user.save
-          format.json  { render :json => user }
-        else
-          format.json  { render :json => 0 }
-        end
-      end
+      user.save
+      response = user
+    end
+
+    respond_to do |format|
+      format.json { render :json => response }
     end
   end
